Save tested working version for evaluation, generation, and parsing

+++ src/pcfg_cpu.cpp
@@ -19,33 +19,33 @@ void pcfg_sample_string_expression_cpu_impl(
     int64_t B)
 {
 #pragma omp parallel for
-    for (int64_t b = 0; b < B; ++b)
+    for (int32_t b = 0; b < B; ++b)
     {
         std::mt19937 gen(seeds[b]);
         std::uniform_real_distribution<float> dis(0.0, 1.0);
         errors[b] = 0;
         bool generated_successfully = false;
-        for (int64_t t = 0; t < max_tries; ++t)
+        for (int32_t t = 0; t < max_tries; ++t)
         {
             bool should_restart = false;
 
-            int64_t stack[HARD_MAX_LENGTH];
-            int64_t stack_ptr = 0;
+            int32_t stack[HARD_MAX_LENGTH];
+            int32_t stack_ptr = 0;
             stack[stack_ptr++] = start_symbol_id;
 
-            int64_t out_ptr = 0;
+            int32_t out_ptr = 0;
 
             while (stack_ptr > 0)
             {
-                int64_t current_symbol = stack[--stack_ptr];
-                int64_t rule_start = nt_rule_ptr[current_symbol];
-                int64_t rule_end = nt_rule_ptr[current_symbol + 1];
+                int32_t current_symbol = stack[--stack_ptr];
+                int32_t rule_start = nt_rule_ptr[current_symbol];
+                int32_t rule_end = nt_rule_ptr[current_symbol + 1];
 
                 if (rule_start != rule_end)
                 { // Non-terminal
                     float rand_val = dis(gen);
-                    int64_t chosen_rule = INVALID_RULE;
-                    for (int64_t r = rule_start; r < rule_end; ++r)
+                    int32_t chosen_rule = INVALID_RULE;
+                    for (int32_t r = rule_start; r < rule_end; ++r)
                     {
                         if (rand_val <= nt_rule_cum_probs[r])
                         {
@@ -60,9 +60,9 @@ void pcfg_sample_string_expression_cpu_impl(
                         break;          // Stop processing this batch item
                     }
 
-                    int64_t rhs_start = rhs_ptr[chosen_rule];
-                    int64_t rhs_end = rhs_ptr[chosen_rule + 1];
-                    for (int64_t i = rhs_end - 1; i >= rhs_start; --i)
+                    int32_t rhs_start = rhs_ptr[chosen_rule];
+                    int32_t rhs_end = rhs_ptr[chosen_rule + 1];
+                    for (int32_t i = rhs_end - 1; i >= rhs_start; --i)
                     {
                         if (stack_ptr >= max_length)
                         {
@@ -97,7 +97,7 @@ void pcfg_sample_string_expression_cpu_impl(
             if (stack_ptr == 0 && out_ptr > 0)
             {
                 // Successfully generated an expression
-                for (int64_t i = out_ptr; i < max_length; ++i)
+                for (int32_t i = out_ptr; i < max_length; ++i)
                 {
                     output[b][i] = NO_OP; // padding
                 }
@@ -123,17 +123,17 @@ void parse_to_postfix_cpu_impl(
     int64_t B, int64_t M)
 {
 #pragma omp parallel for
-    for (int b = 0; b < B; ++b)
+    for (int32_t b = 0; b < B; ++b)
     {
         errors[b] = 0;
-        int64_t op_stack[HARD_MAX_LENGTH];
-        int64_t op_stack_ptr = 0;
-        int64_t out_queue[HARD_MAX_LENGTH];
-        int64_t out_queue_size = 0;
+        int32_t op_stack[HARD_MAX_LENGTH];
+        int32_t op_stack_ptr = 0;
+        int32_t out_queue[HARD_MAX_LENGTH];
+        int32_t out_queue_size = 0;
 
-        for (int i = 0; i < M && expressions_acc[b][i] != NO_OP; ++i)
+        for (int32_t i = 0; i < M && expressions_acc[b][i] != NO_OP; ++i)
         {
-            const int64_t &token = expressions_acc[b][i];
+            const int32_t &token = expressions_acc[b][i];
             if (precedence_acc[token] == 0)
             {
                 if (out_queue_size >= HARD_MAX_LENGTH)
@@ -216,20 +216,20 @@ void parse_to_postfix_cpu_impl(
         if (errors[b] != 0)
             continue;
 
-        for (size_t i = 0; i < out_queue_size; ++i)
+        for (int32_t i = 0; i < out_queue_size; ++i)
         {
             ops_acc[b][i] = out_queue[i];
         }
-        for (size_t i = out_queue_size; i < M; ++i)
+        for (int32_t i = out_queue_size; i < M; ++i)
         {
             ops_acc[b][i] = NO_OP; // padding
         }
 
-        int64_t child_stack[HARD_MAX_LENGTH];
-        int64_t child_stack_ptr = 0;
-        for (size_t i = 0; i < out_queue_size; i++)
+        int32_t child_stack[HARD_MAX_LENGTH];
+        int32_t child_stack_ptr = 0;
+        for (int32_t i = 0; i < out_queue_size; i++)
         {
-            int64_t token = out_queue[i];
+            int32_t token = out_queue[i];
             if (is_unary(token))
             {
                 if (child_stack_ptr < 1)
@@ -237,7 +237,7 @@ void parse_to_postfix_cpu_impl(
                     errors[b] = 5; // Unary operator without operand
                     break;
                 }
-                int64_t child_index = child_stack[--child_stack_ptr];
+                int32_t child_index = child_stack[--child_stack_ptr];
                 children_acc[b][i][0] = child_index;
                 children_acc[b][i][1] = NULL_CHILD;
             }
@@ -248,8 +248,8 @@ void parse_to_postfix_cpu_impl(
                     errors[b] = 6; // Binary operator without enough operands
                     break;
                 }
-                int64_t right_child_index = child_stack[--child_stack_ptr];
-                int64_t left_child_index = child_stack[--child_stack_ptr];
+                int32_t right_child_index = child_stack[--child_stack_ptr];
+                int32_t left_child_index = child_stack[--child_stack_ptr];
                 children_acc[b][i][0] = left_child_index;
                 children_acc[b][i][1] = right_child_index;
             }
